requirements.txt
selene==2.0.0rc9
selenium==4.28.1
pytest~=8.3.4
pytest-asyncio
python-dotenv
webdriver-manager
allure-pytest
allure-python-commons~=2.13.5